Closes #5, Cards within the model grids redirect
Several things were added in this commit. Added a preset for babel to
allow using stage-0 syntax (like the object spread operator, which
is necessary for prop passing). Added Radium as a dependency for
cleaner inline styles. Cleaned up InstanceGrid.js and changed the
InstanceCard to the Card object, and cleaned up Card.js a lot.
Added the inline-styles directory to contain inline styles for our
components to be passed to Radium.